--- Pipfile
@@ -1,15 +1,0 @@
-[[source]]
-name = "pypi"
-url = "https://pypi.org/simple"
-verify_ssl = true
-
-[dev-packages]
-
-[packages]
-plotly = "*"
-flask = "*"
-pandas = "*"
-sklearn = "*"
-
-[requires]
-python_version = "3.6"
