+++ plotly_app/templates/layouts/index.html
